Change the repo_mocking agent factory into standalone function
