+++ ymir/agents/utils.py
@@ -1,5 +1,7 @@
 import logging
 import os
+from collections.abc import Callable
+from typing import Any
 
 from beeai_framework.agents.tool_calling.utils import ToolCallCheckerConfig
 from beeai_framework.backend import ChatModel, ChatModelParameters
@@ -7,6 +9,7 @@ from beeai_framework.template import PromptTemplate
 from pydantic import BaseModel
 
 from ymir.common.base_utils import check_subprocess, run_subprocess  # noqa: F401 — re-exported
+from ymir.common.mock_repos import apply_zstream_override_from_env, setup_mock_repos_from_env
 from ymir.common.utils import get_absolute_path, mcp_tools, run_tool  # noqa: F401 — re-exported
 
 logger = logging.getLogger(__name__)
@@ -88,3 +91,36 @@ def set_litellm_debug() -> None:
     from beeai_framework.adapters.litellm.utils import litellm_debug
 
     litellm_debug(True)
+
+
+def build_agent_factory_with_mock_repos(
+    agent_factory: Callable[[list, dict[str, Any] | None], Any], jira_issue: str
+) -> Callable[[list, dict[str, Any] | None], Any]:
+    """Wrap an agent factory with mock repo support when configured.
+
+    If ``MOCK_REPOS_DIR`` is set, prepares mock repos and returns a factory
+    that injects the resulting ``git_env`` into ``local_tool_options``.
+    Also applies ``MOCK_ZSTREAMS`` (standalone env-var) and any per-issue
+    ``zstream_override`` found in the config file.
+
+    Args:
+        agent_factory: The original agent factory callable
+            (e.g. ``create_triage_agent``).
+        jira_issue: The Jira issue key (e.g. ``RHEL-15216``).
+
+    Returns:
+        The original factory unchanged when mocking is not configured,
+        or a wrapper that injects mock ``git_env``.
+    """
+    apply_zstream_override_from_env()
+
+    git_env = setup_mock_repos_from_env(jira_issue)
+    if git_env is None:
+        return agent_factory
+
+    logger.info("Mock repos configured for %s via MOCK_REPOS_DIR", jira_issue)
+
+    def _factory(gateway_tools, local_tool_options=None):
+        return agent_factory(gateway_tools, {"env": git_env})
+
+    return _factory
